:art: refactor test cases
